pyproject.toml
[project]
name = "cooperation-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "google-genai>=1.48.0",
    "notion-client>=2.7.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
